Fix improper use of the Ruby shovel operator (<<)
junk would be modified and returned, and we just want to return the
concatenated string. Practically doesn't matter, but it's incorrect.

This was my first public module. I've been wanting to fix this since.
I'm noticing it again now as I look for how I used Ret in a target.