Add Webmin 1900 Remote Code Exec Module
Adding Webmin RCE module affecting Webmin <= 1.900. Module attempts to
use the Running Processes (proc) permission to determine upload dir,
if the permission is not set the module fails. The user can attempt the
exploit without this permission by setting the 'GUESSUPLOAD' opt to
true.

The default path is in an array of 1 to allow for other OS/Version
default paths to be added in future.