Fix bug in tab completion in case-sensitive situations
Includes a fix for a crash caused when tab completing.
This crash is the result of a bug in rb-readline when the only remaining options
provided in the tab completion procedure are identical. The bug is only triggered
when going case-insensitive, but we can avoid it by de-duping the list sent back to
rb-readline.